+++ src/main/java/de/maxhenkel/plane/DamageSourcePlane.java
@@ -7,6 +7,7 @@ import net.minecraft.world.damagesource.DamageType;
 
 public class DamageSourcePlane {
 
-    public static final ResourceKey<DamageType> DAMAGE_PLANE_TYPE = ResourceKey.create(Registries.DAMAGE_TYPE, ResourceLocation.fromNamespaceAndPath(Main.MODID, "hit_plane"));
+    public static final ResourceKey<DamageType> DAMAGE_HIT_PLANE = ResourceKey.create(Registries.DAMAGE_TYPE, ResourceLocation.fromNamespaceAndPath(Main.MODID, "hit_plane"));
+    public static final ResourceKey<DamageType> DAMAGE_PLANE_CRASH = ResourceKey.create(Registries.DAMAGE_TYPE, ResourceLocation.fromNamespaceAndPath(Main.MODID, "plane_crash"));
 
 }