+++ changelog.md
@@ -1,1 +1,2 @@
 - Fixed players getting damaged when jumping out of planes
+- Damage players when crashing a plane